Move enhanced_library from helpers to utils
