+++ google/cloud/aiplatform/utils/enhanced_library/__init__.py
@@ -1,4 +1,4 @@
-# Copyright 2021 Google LLC
+# Copyright 2020 Google LLC
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -11,9 +11,3 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
-from google.cloud.aiplatform.helpers.container_uri_builders import (
-    get_prebuilt_prediction_container_uri,
-)
-
-__all__ = "get_prebuilt_prediction_container_uri"
