+++ tests/unit/enhanced_library/test_value_converter.py
@@ -13,7 +13,7 @@
 # limitations under the License.
 from __future__ import absolute_import
 
-from google.cloud.aiplatform.helpers import value_converter
+from google.cloud.aiplatform.utils.enhanced_library import value_converter
 from google.protobuf import json_format
 from google.protobuf.struct_pb2 import Value
 import proto
